Move RemoteFeedImageDataLoader to production
